[server] Create new sub resources for all current api resources
A sub resource is very light-weight version of an api resource and is meant to give the bare minimum details about the resource.

From now on, api resources should only link to sub resources, and sub resources should not link to either api resources or other sub resources. This should keep the depth of objects returned from the api to 2 levels deep only. This keeps thing simple for now, but we can adjust this later if we encounter a requirement to go further than 2 levels deep.