[server] Validate Sprint for newly created Backlog Items
The sprint should belong to the same project as the backlog item to be created.

Note: validation for existing backlog items still needs to be done, but it will require updating the BacklogItemPatchDto to use a backlogItemId. This is required in order to determine which project the backlog item being patched belongs to.