Remove ProjectRepository/Repositories and create new Models + Converters for the application layer
There really isn't a need for repositories in general as DBContext takes care of most repository needs.

Creating new models specifically for the application layer to use, so that entities no longer need to be used at the application layer.

Creating new converters to convert entities to models and vice versa.