[server] Update Model Hydrators to use a hydrator registry to reduce dependencies between hydrators
When a hydrator needs to hydrate an object that it is not responsible for hydrating, it asks the registry to hydrate the object, so that the hydrator doesn't need to know or care about other hydrators. This should help to reduce dependencies between them and to solve cyclic dependency injection issues.