[server] Generalize Hydrator logic and move it out of the application layer.
Update HydratorRegistry to depend on IEnumerable<IHydrator>, so we don't need to manually load specific hydrators via HydratorLoader.

Update Hydrate() interface to include a new argument for a referenceHydrator, which can be used to hydrate references as needed.

Create a general Hydrator(), which can be used to perform all hydrations.